functions removed from main script, utils added
