+++ src/create_blast_db.py
@@ -6,7 +6,7 @@ import json
 import logging
 import sys
 from datetime import datetime
-from ftplib import FTP
+
 from pathlib import Path
 from shutil import copyfile
 from subprocess import PIPE, Popen
@@ -16,42 +16,12 @@ import wget
 from rich.console import Console
 import yaml
 
+from utils import check_md5sum, extendable_logger, get_ftp_file_size, get_mod_from_json
+
+
 console = Console()
 MAKEBLASTDB_BIN = "/usr/local/bin/makeblastdb"
 MODS = ["FB", "SGD", "WB", "XB", "ZFIN"]
-
-
-def extendable_logger(log_name, file_name, level=logging.INFO):
-    """
-    Function that creates a logger that can be extended
-    :param log_name:
-    :param file_name:
-    :param level:
-    """
-    formatter = logging.Formatter("[%(asctime)s] %(levelname)s %(message)s")
-    handler = logging.FileHandler(file_name)
-    handler.setFormatter(formatter)
-    specified_logger = logging.getLogger(log_name)
-    specified_logger.setLevel(level)
-    specified_logger.addHandler(handler)
-
-    return specified_logger
-
-
-def check_md5sum(fasta_file, md5sum) -> bool:
-    """
-    Function that checks the md5sum of the downloaded file
-    :param fasta_file:
-    :param md5sum:
-    """
-
-    downloaded_md5sum = hashlib.md5(open(fasta_file, "rb").read()).hexdigest()
-    if downloaded_md5sum != md5sum:
-        console.log(f"MD5sums do not match: {md5sum} != {downloaded_md5sum}")
-        return False
-    else:
-        console.log(f"MD5sums match: {md5sum} {downloaded_md5sum}")
-        return True
 
 
 def store_fasta_files(fasta_file, file_logger) -> None:
@@ -205,22 +175,7 @@ def run_makeblastdb(config_entry, output_dir, file_logger):
     return True
 
 
-def get_mod_from_json(input_json) -> str:
-    """
-    Function that gets the mod from the JSON file
-    :param input_json:
-    """
 
-    filename = Path(input_json).name
-    mod = filename.split(".")[1]
-
-    if mod not in MODS:
-        console.log(f"Mod {mod} not found in {MODS}")
-        return False
-
-    console.log(f"Mod found: {mod}")
-
-    return mod
 
 
 def process_yaml(config_yaml) -> bool:
